Add rollcontrol authentication plugin
RollControl is a plugin that manages authorization based on user email
address. It allows for users to have rights to run commands on certain
repositories, etc without giving global permissions for all slack users
to run all commands, and without giving a particular user with the right
to one repository the rights on another.